Consolidated unit tests for upwind discretization
